requirements.txt
dj-database-url==0.5.0
Django==2.1.5
django-jsonfield==1.0.1
django-notify-x==0.1.9
django-summernote==0.8.11.4
gunicorn==19.9.0
psycopg2-binary==2.7.7
pytz==2018.9
six==1.12.0
whitenoise==4.1.2
